TeamCity: refactor FunctionalTestBucket code to move shared types to a separate file
