+++ .teamcity/src/main/kotlin/model/FunctionalTestBucketProvider.kt
@@ -198,47 +198,3 @@ class GradleVersionRangeCrossVersionTestBucket(
         )
     }
 }
-
-data class SmallSubprojectBucket(
-    val subprojects: List<GradleSubproject>,
-    val parallelizationMethod: ParallelizationMethod,
-) : BuildTypeBucket {
-    constructor(
-        subproject: GradleSubproject,
-        parallelizationMethod: ParallelizationMethod,
-    ) : this(listOf(subproject), parallelizationMethod)
-
-    val name = truncateName(subprojects.joinToString(","))
-
-    private fun truncateName(str: String) =
-        // Can't exceed Linux file name limit 255 char on TeamCity
-        if (str.length > 200) {
-            str.substring(0, 200) + "..."
-        } else {
-            str
-        }
-
-    override fun createFunctionalTestsFor(
-        model: CIBuildModel,
-        stage: Stage,
-        testCoverage: TestCoverage,
-        bucketIndex: Int,
-    ): FunctionalTest =
-        FunctionalTest(
-            model,
-            testCoverage.getBucketUuid(model, bucketIndex),
-            getName(testCoverage),
-            getDescription(testCoverage),
-            testCoverage,
-            stage,
-            parallelizationMethod,
-            subprojects.map { it.name },
-        )
-
-    override fun getName(testCoverage: TestCoverage) =
-        truncateName("${testCoverage.asName()} (${subprojects.joinToString(",") { it.name }})")
-
-    override fun getDescription(testCoverage: TestCoverage) = "${testCoverage.asName()} for ${subprojects.joinToString(", ") { it.name }}"
-
-    fun toJsonBucket(): FunctionalTestBucket = FunctionalTestBucket(subprojects.map { it.name }, parallelizationMethod)
-}
